+++ tests/consent_mailer/test_extract.py
@@ -17,8 +17,12 @@ def test_extract_ldap_persons(
 
 
 @pytest.mark.usefixtures("mocked_consent_backend_api_connector")
-def test_extract_consents_for_persons() -> None:
-    # due to its mocked, we dont need any param
+def test_extract_consents_for_persons(
+    mocked_consent_backend_api_connector: dict[str, int],
+) -> None:
     persons = extract_ldap_persons(MergedPrimarySourceIdentifier(""))
     consents = extract_consents_for_persons(persons)
     assert len(consents) == 1
+    # assert 3 calls: 1 call for 2 persons, 2 calls for consent (filter for
+    # 2 persons in batches of size 1)
+    assert mocked_consent_backend_api_connector["count"] == 3